third_party/blink/web_tests/platform/linux-chrome/external/wpt/permissions/permissions-query-feature-policy-attribute.https.sub-expected.txt

This is a testharness.js-based test.
[PASS] Permissions.state is "prompt" with allow="geolocation" in same-origin iframes.
[PASS] Permissions.state is "prompt" with allow="geolocation" in cross-origin iframes.
[PASS] Permission.state is "prompt" in same-origin iframes.
[FAIL] Permission.state is "denied" in cross-origin iframes.
  assert_equals: navigator.permissions.query("geolocation") expected "denied" but got "prompt"
Harness: the test ran to completion.
